NODE-1283: Move all DownloadManager related code under the new package 'io.casperlabs.comm.gossiping.downloadmanager' without any semantic modifications
