+++ comm/src/main/scala/io/casperlabs/comm/gossiping/GenesisApprover.scala
@@ -1,10 +1,9 @@
 package io.casperlabs.comm.gossiping
 
-import cats._
+import cats.data.NonEmptyList
 import cats.effect._
 import cats.effect.concurrent._
 import cats.implicits._
-import cats.syntax._
 import com.google.protobuf.ByteString
 import io.casperlabs.casper.consensus._
 import io.casperlabs.comm.ServiceError
@@ -12,12 +11,12 @@ import io.casperlabs.comm.ServiceError.{Internal, InvalidArgument, NotFound, Una
 import io.casperlabs.comm.discovery.NodeUtils._
 import io.casperlabs.comm.discovery.{Node, NodeDiscovery}
 import io.casperlabs.comm.gossiping.Utils._
+import io.casperlabs.comm.gossiping.downloadmanager.BlockDownloadManager
 import io.casperlabs.shared.Log
 
-import scala.concurrent.duration.{Duration, FiniteDuration}
+import scala.concurrent.duration.FiniteDuration
 import scala.util.Random
 import scala.util.control.NonFatal
-import cats.data.NonEmptyList
 
 /** Accumulate approvals for the Genesis block. When enough of them is
   * present to pass a threshold which is the preorgative of this node,
