+++ comm/src/test/scala/io/casperlabs/comm/gossiping/BlockDownloadManagerSpec.scala
@@ -10,7 +10,8 @@ import io.casperlabs.casper.consensus.{Approval, Block, BlockSummary}
 import io.casperlabs.comm.GossipError
 import io.casperlabs.comm.discovery.Node
 import io.casperlabs.comm.discovery.NodeUtils.showNode
-import io.casperlabs.comm.gossiping.BlockDownloadManagerImpl.RetriesConf
+import io.casperlabs.comm.gossiping.downloadmanager.BlockDownloadManagerImpl._
+import io.casperlabs.comm.gossiping.downloadmanager._
 import io.casperlabs.comm.gossiping.synchronization.Synchronizer
 import io.casperlabs.metrics.Metrics
 import io.casperlabs.models.BlockImplicits.BlockOps
